refactor(gda-balancing): own operation evidence replay
